Upodate v_03
Battle System - for now it has only 5 monsters.
Enemy Duel pop up - fight/run ( i will finish the run method in the next update)
UI text which tells you whe nyou are ready for duel and when ou are not and other small things

+++ TapAdventureGame/Assets/Scripts/Character.cs
@@ -7,5 +7,5 @@ public class Character {
 
     public float Damage = 5;
     public float Health= 8;
-    public int Level = 1;
+    public float Level = 1;
 }
